Задача 1. Иправлены мапперы, разделены модели на разные файлы, удалено свойство type в data моделях
