+++ app/src/main/java/com/example/test_app/data/repositories/DentistryProgramRepositoryImpl.kt
@@ -1,8 +1,7 @@
 package com.example.test_app.data.repositories
 
 import com.example.test_app.data.api.ApiHelper
-import com.example.test_app.data.mappers.DentistryProgramMapper
-import com.example.test_app.data.models.DentistryProgramApiDto
+import com.example.test_app.data.mappers.toDomain
 import com.example.test_app.domain.models.DentistryProgram
 import com.example.test_app.domain.repositories.DentistryProgramRepository
 import com.example.test_app.util.ExceptionHandler.runWithTryCatch
@@ -20,9 +19,7 @@ class DentistryProgramRepositoryImpl : DentistryProgramRepository {
             ApiHelper.apiClient.getDentistryProgram(dentistryId).await()
         }
 
-        return DentistryProgramMapper.toDomain(
-            result ?: DentistryProgramApiDto(description = "", actions = emptyList())
-        )
+        return result?.toDomain() ?: DentistryProgram(description = "", actions = emptyList())
     }
 
 }
